requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
python-dotenv
pytest
